Tests reconstruction
